bash-ops/requirements.txt
GitPython
pyyaml
paramiko
ansible==2.10.7
tabulate
